Dev ex improvements (#130)
* Reorganize contracts

* Preliminary linting

* Fix linting issues

* Prettier setup for Solidity

* Remove old src files

* Prettier fix

* Linting migrations, remove dead code

* Remove merged contracts

* Convert scripts to typescript

* Deploy otoken

* Fix test

* test

* test again

* Split out into further lints

* Fix lint

* Add new check

* Actions

* Lint sol

* off error

* Clean up files

* Add back MockERC20

* Update README.md

+++ contracts/interfaces/ISwapPair.sol
@@ -3,6 +3,19 @@ pragma solidity >=0.7.2;
 pragma experimental ABIEncoderV2;
 
 interface ISwapPair {
-    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+    function getReserves()
+        external
+        view
+        returns (
+            uint112 reserve0,
+            uint112 reserve1,
+            uint32 blockTimestampLast
+        );
+
+    function swap(
+        uint256 amount0Out,
+        uint256 amount1Out,
+        address to,
+        bytes calldata data
+    ) external;
 }
